Fixed the bugs in the laps

+++ neuropy/core/neurons.py
@@ -497,9 +497,8 @@ class Neurons(NeuronUnitSlicableObjectProtocol, StartStopTimesMixin, TimeSlicabl
         for neuron_idx in np.arange(num_neurons):
             for obj_idx in np.arange(1, len(objList)):
                 # spiketrains_list[neuron_idx].append(objList[obj_idx].spiketrains[neuron_idx])
-                spiketrains_list[neuron_idx] = np.append(spiketrains_list[neuron_idx], objList[obj_idx].spiketrains[neuron_idx])
+                spiketrains_list[neuron_idx] = np.append(spiketrains_list[neuron_idx], objList[obj_idx].spiketrains[neuron_idx]).astype(np.float64)
                 
-        
             # for obj_idx in np.arange(len(objList)):
             # # spiketrains_list[i] =  np.concatenate([obj.spiketrains for obj in objList], axis=0)
             # spiketrains_list.append(np.concatenate([obj.spiketrains[neuron_idx] for neuron_idx in np.arange(num_neurons)], axis=0))
